Fix date substitution in AGENTS.md for new sessions (#32363)
- Add formatDateStampInTimezone to date-time.ts
- Add date substitution in loadWorkspaceBootstrapFiles
- Pass config through bootstrap files pipeline
- Add date substitution in post-compaction context
- Remove duplicate function definition in memory-flush.ts

Fixes #32363
